+++ packages/gamut/src/List/__tests__/List-test.tsx
@@ -14,22 +14,22 @@ const renderWrapper = setupEnzyme(List, {
 });
 
 describe('Menu', () => {
-  it('renders a slat list by default', () => {
+  it('renders a default list by default', () => {
     const { wrapper } = renderWrapper();
 
-    expect(wrapper.find('ListEl').prop('variant')).toBe('slat');
-    expect(wrapper.find('ListEl').prop('scrollable')).toBe(false);
+    expect(wrapper.find('ListEl').prop('variant')).toBe('default');
+    expect(wrapper.find('ListEl').prop('scrollable')).toBe(undefined);
   });
   it('configures rows with the correct variants', () => {
     const { wrapper } = renderWrapper();
-
-    expect(wrapper.find('RowEl').prop('variant')).toBe('slat');
-    expect(wrapper.find('RowEl').prop('spacing')).toBe('normal');
+    const wrappingRow = wrapper.find('RowEl').at(0);
+    expect(wrappingRow.prop('variant')).toBe('default');
+    expect(wrappingRow.prop('spacing')).toBe('normal');
   });
   it('configures columns with the correct variants', () => {
     const { wrapper } = renderWrapper();
 
-    expect(wrapper.find('ColEl').prop('variant')).toBe('slat');
+    expect(wrapper.find('ColEl').prop('variant')).toBe('default');
     expect(wrapper.find('ColEl').prop('spacing')).toBe('normal');
     expect(wrapper.find('ColEl').prop('sticky')).toBe(false);
   });