+++ packages/gamut-labs/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [25.4.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.4.0...@codecademy/gamut-labs@25.4.1) (2021-10-21)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+## [25.4.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.3.0...@codecademy/gamut-labs@25.4.0) (2021-10-20)
+
+
+### Features
+
+* **GlobalHeader:** [GM-399] Adding dark mode to GlobalHeader and all child components ([1148a31](https://github.com/Codecademy/client-modules/commit/1148a31ffb4fb320f3a5d1c6a9669469a3084712))
+
+
+
+## [25.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.2.2...@codecademy/gamut-labs@25.3.0) (2021-10-20)
+
+
+### Features
+
+* **PageAlerts:** porting PageAlerts from Portal-App ([9ab0dc1](https://github.com/Codecademy/client-modules/commit/9ab0dc1d8d15e3534fb85267abcf7430e6528665))
+
+
+
+### [25.2.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.2.1...@codecademy/gamut-labs@25.2.2) (2021-10-20)
+
+
+### Bug Fixes
+
+* **Drawer:** use minWidth instead of width in internal Box ([d19d483](https://github.com/Codecademy/client-modules/commit/d19d4836124f5e17b241813978dbb954f02d260d))
+
+
+
+### [25.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.2.0...@codecademy/gamut-labs@25.2.1) (2021-10-19)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
 ## [25.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@25.1.8...@codecademy/gamut-labs@25.2.0) (2021-10-15)